Added delete endpoint
